refactor LottoJudge.java: 수익률을 double로 리턴하도록 변경
수익률을 double로 리턴하고 resultView에서 소수점 둘째짜리까지 출력하도록 함

+++ src/main/java/lotto/dto/LottoStatisticsDTO.java
@@ -3,19 +3,19 @@ package lotto.dto;
 import java.util.Objects;
 
 public class LottoStatisticsDTO {
-    private final int percentageRateOfReturn;
+    private final double percentageRateOfReturn;
     private final LottoMatchInfosDTO matchInfosDTO;
 
-    private LottoStatisticsDTO(int percentageRateOfReturn, LottoMatchInfosDTO matchInfosDTO) {
+    private LottoStatisticsDTO(double percentageRateOfReturn, LottoMatchInfosDTO matchInfosDTO) {
         this.percentageRateOfReturn = percentageRateOfReturn;
         this.matchInfosDTO = matchInfosDTO;
     }
 
-    public static LottoStatisticsDTO valueOf(int percentageRateOfReturn, LottoMatchInfosDTO matchCountsDTO) {
+    public static LottoStatisticsDTO valueOf(double percentageRateOfReturn, LottoMatchInfosDTO matchCountsDTO) {
         return new LottoStatisticsDTO(percentageRateOfReturn, matchCountsDTO);
     }
 
-    public int getPercentageRateOfReturn() {
+    public double getPercentageRateOfReturn() {
         return percentageRateOfReturn;
     }
